map group items have hex start/end coords
