+++ src/main/java/de/bonndan/nivio/output/map/GroupMapItem.java
@@ -17,8 +17,8 @@ class GroupMapItem extends MapItem {
     public final Group group;
     public final long x;
     public final long y;
-    public final double width;
-    public final double height;
+    public final long width;
+    public final long height;
 
     public GroupMapItem(Group group, long x, long y, long width, long height) {
         super(group.getIdentifier(), group.getIdentifier(), "", "group", Color.getGroupColor(group));
@@ -28,5 +28,13 @@ class GroupMapItem extends MapItem {
         this.width = width;
         this.height = height;
         this.group = group;
+    }
+
+    public Hex getStart() {
+        return asHex(x, y, size);
+    }
+
+    public Hex getEnd() {
+        return asHex(x + width, y + height, size);
     }
 }
