+++ src/main/java/de/bonndan/nivio/output/map/ItemMapItem.java
@@ -10,8 +10,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static java.lang.Math.sqrt;
-
 /**
  * JSON representation for custom rendering.
  * <p>
@@ -25,7 +23,6 @@ class ItemMapItem extends MapItem {
     public long y;
     public double width;
     public double height;
-    public int size;
     public final String group;
     public final LandscapeItem landscapeItem;
     public final List<String> relations = new ArrayList<>();
@@ -46,17 +43,8 @@ class ItemMapItem extends MapItem {
         relations.addAll(collect);
     }
 
-    /**
-     * Calculated hexagon map coordinates from x-y coordinates.
-     * <p>
-     * https://stackoverflow.com/questions/20734438/algorithm-to-generate-a-hexagonal-grid-with-coordinate-system/20751975#20751975
-     * https://www.redblobgames.com/grids/hexagons/#rounding
-     */
     public Hex getHex() {
-        var q = (2. / 3 * x) / size;
-        var r = (-1. / 3 * x + sqrt(3) / 3 * y) / size;
-        FractionalHex hex = new FractionalHex(q, r, -q - r);
-        return hex.hexRound();
+        return asHex(x, y, size);
     }
 
 }
